LogicOverhaul: Use conftest for autouse fixtures
