+++ test/test_import_tracker.py
@@ -14,7 +14,6 @@ import warnings
 import pytest
 
 # Local
-from test.helpers import remove_test_deps, reset_sys_modules
 import import_tracker
 
 ## Package API #################################################################
